refactor: improve zone typing and fix casing
