+++ src/components/icon/Shlagidiamond.vue
