+++ src/data/zones/villages/village50.ts
@@ -1,4 +1,4 @@
-import type { Zone } from '~/type'
+import type { SavageZoneId, Zone } from '~/type'
 import { VILLAGE_OFFSET } from '~/constants/zone'
 import {
   attackPotion,
@@ -26,7 +26,7 @@ export const village50: Zone = {
   type: 'village',
   villageType: 'super',
   position: move.top(savage45.position, VILLAGE_OFFSET),
-  attachedTo: savage45.id,
+  attachedTo: savage45.id as SavageZoneId,
   actions: [],
   minLevel: 50,
   village: {
