+++ src/data/zones/villages/village60.ts
@@ -1,4 +1,4 @@
-import type { Zone } from '~/type'
+import type { SavageZoneId, Zone } from '~/type'
 import { VILLAGE_OFFSET } from '~/constants/zone'
 import {
   attackPotion,
@@ -28,7 +28,7 @@ export const village60: Zone = {
   type: 'village',
   villageType: 'hyper',
   position: move.top(savage55.position, VILLAGE_OFFSET),
-  attachedTo: savage55.id,
+  attachedTo: savage55.id as SavageZoneId,
   actions: [
     { id: 'minigame', label: 'Mini-jeu' },
   ],
